work on FoodChoicesView for host finished

+++ app/src/main/java/com/example/boardgamer/EssenWahl.java
@@ -5,5 +5,5 @@ import com.google.gson.annotations.SerializedName;
 public class EssenWahl {
     @SerializedName("spieltermin_id") public long spieltermin_id;
     @SerializedName("spieler_name") public String spieler_name;
-    @SerializedName("essensrichtung_id") public int essensrichtung_id;
+    @SerializedName("essensrichtung_id") public long essensrichtung_id;
 }
